backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.1.0
django-environ==0.11.2
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.8
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
